moved Light stuff into topics
